adaptation of cameramap
